Ports config tests for Postfix to separate file
We have a number of Postfix-specific checks in the OSSEC testinfra
suite, so let's break them into a separate file before adding more.
Looks like we never added an "Is Postfix running?" check, will do that
shortly.
